requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
email-validator
pandas==2.1.1
pytest==7.4.2
pytest-cov==4.1.0
gunicorn==21.2.0
psycopg2-binary
Flask-Script==2.0.6
Faker==19.3.0
Pillow==10.1.0
matplotlib
seaborn